+++ api/V0ComposeUpdate.go
@@ -51,6 +51,11 @@ func (e *API) V0ComposeUpdate(w http.ResponseWriter, r *http.Request) {
 			key := e.Redis.GetRedisKey(keys.Val())
 			task := mesosutil.DecodeTask(key)
 			e.mapComposeServiceToMesosTask(vars, service, task)
+
+			// restore the old MesosAgent info
+			updatedTask := mesosutil.DecodeTask(key)
+			updatedTask.MesosAgent = task.MesosAgent
+			e.Redis.SaveTaskRedis(updatedTask)
 		}
 	}