+++ changelog.md
@@ -27,6 +27,8 @@
 - FIX: restart services and tasks to prevent unmanaged tasks
 - ADD: Support for user defined network with exposed ports
 - ADD: customize taskname `bis.aventer.mesos_compose.taskname: "test:app"`. 
+- FIX: restore MesosAgent info after update task by API
+- FIX: Remove LOST mesos tasks from redis
   
 ## v0.4.0